llm_calling important updates for function
